chore: deprecates @dfinity/authentication (#661)
* deprecates @dfinity/authentication and moves isDelegationValid check to @dfinity/identity

* deprecates identity-ledgerhq

* removes name field from jest configs

* deprecation instructions

* removes ledgerhq demo
