--- demos/ledgerhq/README.md
@@ -1,9 +1,0 @@
-# ledgerhq, agent-js
-
-## Development
-
-### Getting Started
-
-1. Ensure all dependencies are installed: `npm install`
-2. Run the development server `npm run develop`
-3. Visit the running site at http://localhost:8080
