--- demos/ledgerhq/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "ic-agent-ledgerhq-app",
-  "private": true,
-  "dependencies": {
-    "@dfinity/agent": "^0.14.1",
-    "@dfinity/authentication": "^0.14.1",
-    "@dfinity/identity": "^0.14.1",
-    "@dfinity/identity-ledgerhq": "^0.14.1",
-    "@dfinity/principal": "^0.14.1",
-    "assert": "^2.0.0",
-    "buffer": "^6.0.3",
-    "events": "^3.2.0",
-    "html-webpack-plugin": "^5.1.0",
-    "process": "^0.11.10",
-    "protobufjs": "^6.10.2",
-    "stream-browserify": "^3.0.0",
-    "ts-node": "^10.8.2",
-    "util": "^0.12.3",
-    "webpack": "^5.24.1",
-    "webpack-cli": "^4.5.0",
-    "webpack-dev-server": "^4.0.0-beta.0"
-  },
-  "scripts": {
-    "build": "webpack",
-    "develop": "webpack-dev-server",
-    "lint:fix": "npm run lint -- --fix",
-    "publish:release": "",
-    "test:coverage": "",
-    "test": ""
-  },
-  "version": "0.14.1",
-  "devDependencies": {
-    "esbuild": "^0.15.16",
-    "size-limit": "^8.1.0"
-  }
-}
