--- demos/ledgerhq/webpack.config.js
@@ -1,57 +1,0 @@
-const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const webpack = require('webpack');
-
-module.exports = {
-  entry: {
-    main: path.join(__dirname, 'src/main.js'),
-  },
-  mode: 'production',
-  target: 'web',
-  output: {
-    path: path.join(__dirname, 'dist'),
-    filename: '[name].bundle.js',
-  },
-  optimization: {
-    splitChunks: {
-      chunks: 'all',
-      minSize: 20000,
-      maxSize: 200000,
-      minRemainingSize: 0,
-      minChunks: 1,
-      maxAsyncRequests: 30,
-      maxInitialRequests: 30,
-      enforceSizeThreshold: 50000,
-      cacheGroups: {
-        default: {
-          minChunks: 2,
-          priority: -20,
-          reuseExistingChunk: true,
-        },
-      },
-    },
-  },
-  resolve: {
-    alias: {
-      process: 'process/browser',
-    },
-    fallback: {
-      assert: require.resolve('assert/'),
-      buffer: require.resolve('buffer/'),
-      events: require.resolve('events/'),
-      stream: require.resolve('stream-browserify/'),
-      util: require.resolve('util/'),
-    },
-  },
-  devtool: 'source-map',
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: 'src/index.html',
-      filename: 'index.html',
-    }),
-    new webpack.ProvidePlugin({
-      Buffer: [require.resolve('buffer/'), 'Buffer'],
-      process: require.resolve('process/browser'),
-    }),
-  ],
-};
