+++ demos/sample-javascript/webpack.config.js
@@ -43,6 +43,7 @@ module.exports = {
     },
     fallback: {
       assert: require.resolve('assert/'),
+      buffer: require.resolve('buffer/'),
       events: require.resolve('events/'),
       stream: require.resolve('stream-browserify/'),
       util: require.resolve('util/'),
