+++ packages/authentication/README.md
@@ -1,5 +1,8 @@
 # @dfinity/authentication
 
+**Warning**
+this package is deprecated
+
 JavaScript and TypeScript library to support manage Identities and enable simple Web Authentication flows for applications on the [Internet Computer](https://dfinity.org/)
 
 Visit the [Dfinity Forum](https://forum.dfinity.org/) and [SDK Documentation](https://sdk.dfinity.org/docs/index.html) for more information and support building on the Internet Computer.
