+++ packages/authentication/package.json
@@ -4,79 +4,13 @@
   "author": "DFINITY Stiftung <sdk@dfinity.org>",
   "license": "Apache-2.0",
   "description": "JavaScript and TypeScript library to manage identity and authentication with the Internet Computer",
-  "homepage": "https://smartcontracts.org",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/dfinity/agent-js.git",
-    "directory": "packages/authentication"
-  },
-  "bugs": {
-    "url": "https://github.com/dfinity/agent-js/issues"
-  },
-  "keywords": [
-    "internet computer",
-    "ic",
-    "dfinity",
-    "canister",
-    "webauthn",
-    "identity",
-    "principal",
-    "authentication",
-    "dfx",
-    "candid",
-    "motoko",
-    "javascript",
-    "typescript",
-    "blockchain",
-    "crypto",
-    "distributed",
-    "api"
-  ],
+  "deprecated": "This package has been deprecated. isDelegationValid has been migrated to @dfinity/identity",
   "main": "./lib/cjs/index.js",
-  "module": "./lib/esm/index.js",
-  "unpkg": "./lib/esm/index",
   "scripts": {
-    "build": "tsc -b && tsc -p tsconfig-cjs.json",
-    "bundle": "esbuild --bundle src/index.ts --outfile=dist/index.js",
-    "size-limit": "size-limit",
-    "lint": "eslint 'src' --ext '.js,.jsx,.ts,.tsx'",
-    "lint:fix": "npm run lint -- --fix",
-    "make:docs/reference": "typedoc src/index.ts --out ../../docs/generated/authentication --excludeInternal",
-    "release": "npm publish",
-    "test": "jest",
-    "test:coverage": "jest --verbose --collectCoverage"
-  },
-  "peerDependencies": {
-    "@dfinity/agent": "^0.14.1",
-    "@dfinity/identity": "^0.14.1",
-    "@dfinity/principal": "^0.14.1"
-  },
-  "devDependencies": {
-    "@trust/webcrypto": "^0.9.2",
-    "@types/jest": "^28.1.4",
-    "@typescript-eslint/eslint-plugin": "^5.30.5",
-    "@typescript-eslint/parser": "^5.30.5",
-    "esbuild": "^0.15.16",
-    "eslint": "^8.19.0",
-    "eslint-plugin-jsdoc": "^39.3.3",
-    "jest": "^28.1.2",
-    "size-limit": "^8.1.0",
-    "text-encoding": "^0.7.0",
-    "ts-jest": "^28.0.5",
-    "ts-node": "^10.8.2",
-    "tslint": "^5.20.0",
-    "typedoc": "^0.22.11",
-    "typescript": "^4.7.4",
-    "whatwg-fetch": "^3.0.0"
+    "build": "",
+    "test": ""
   },
   "dependencies": {
-    "stream": "^0.0.2"
-  },
-  "size-limit": [
-    {
-      "path": "./dist/index.js",
-      "limit": "380 kB",
-      "webpack": false
-    }
-  ]
+    "@dfinity/identity": "^0.14.1"
+  }
 }
