--- packages/authentication/test-setup.ts
@@ -1,14 +1,0 @@
-// This file may be used to polyfill features that aren't available in the test
-// environment, i.e. JSDom.
-//
-// We sometimes need to do this because our target browsers are expected to have
-// a feature that JSDom doesn't.
-//
-// Note that we can use webpack configuration to make some features available to
-// Node.js in a similar way.
-
-global.crypto = require('@trust/webcrypto');
-global.TextEncoder = require('text-encoding').TextEncoder;
-global.TextDecoder = require('text-encoding').TextDecoder;
-global.MessageChannel = require('worker_threads').MessageChannel;
-require('whatwg-fetch');
