--- packages/authentication/tsconfig.json
@@ -1,23 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "allowJs": false,
-    "baseUrl": "./",
-    "composite": true,
-    "declaration": true,
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "incremental": true,
-    "module": "ES2020",
-    "moduleResolution": "node",
-    "outDir": "./lib/esm",
-    "resolveJsonModule": true,
-    "rootDir": "./src",
-    "sourceMap": true,
-    "inlineSources": true,
-    "strict": true,
-    "target": "es2017"
-  },
-  "include": ["types/*", "src/**/*"],
-  "references": [{ "path": "../agent" }, { "path": "../identity" }]
-}
