--- packages/authentication/tslint.json
@@ -1,24 +1,0 @@
-{
-  "defaultSeverity": "error",
-  "extends": ["tslint:recommended"],
-  "jsRules": {},
-  "linterOptions": {
-    "exclude": ["src/IDL.js", "src/IDL.test.js"]
-  },
-  "rules": {
-    "max-classes-per-file": false,
-    "interface-name": [true, "never-prefix"],
-    "no-consecutive-blank-lines": [true, 2],
-    "no-empty": [true, "allow-empty-functions"],
-    "no-switch-case-fall-through": true,
-    "object-literal-sort-keys": false,
-    "max-line-length": [true, 100],
-    "quotemark": [true, "single"],
-    "arrow-parens": [true, "ban-single-arg-parens"],
-    "space-before-function-paren": [false],
-    "variable-name": {
-      "options": ["ban-keywords", "check-format", "allow-leading-underscore", "allow-pascal-case"]
-    }
-  },
-  "rulesDirectory": []
-}
