--- packages/identity-ledgerhq/.gitignore
@@ -1,17 +1,0 @@
-build_info.json
-node_modules/
-dist/
-lib/**/*.js
-lib/**/*.js.map
-lib/**/*.d.ts
-
-# generated docs
-/docs/reference
-
-# Cannot ignore .d.ts files in types/
-!types/**/*.d.ts
-
-# Cannot ignore setup files for webpack and jest, which are still JavaScript.
-!webpack.config.js
-!jest.config.js
-!test-setup.js
