--- packages/identity-ledgerhq/src/identity/ledger_hq__hw_transport_webhid.d.ts
@@ -1,1 +1,0 @@
-declare module '@ledgerhq/hw-transport-webhid';
