--- packages/identity-ledgerhq/src/index.ts
@@ -1,2 +1,0 @@
-export * from './identity/ledger';
-export * from './identity/secp256k1';
