+++ packages/identity/src/index.ts
@@ -1,11 +1,6 @@
 export { Ed25519KeyIdentity, Ed25519PublicKey } from './identity/ed25519';
 export * from './identity/ecdsa';
 export { Secp256k1KeyIdentity, Secp256k1PublicKey } from './identity/secp256k1';
-export {
-  Delegation,
-  DelegationIdentity,
-  DelegationChain,
-  SignedDelegation,
-} from './identity/delegation';
+export * from './identity/delegation';
 export { WebAuthnIdentity } from './identity/webauthn';
 export { wrapDER, unwrapDER, DER_COSE_OID, ED25519_OID } from './identity/der';
